add major crud

+++ routes/web.php
@@ -18,12 +18,30 @@ use Illuminate\Support\Facades\Route;
 Route::get('/', function () {
     return view('front.home');
 });
+Route::get('/majors', function () {
+    return view('front.pages.majors');
+});
+Route::get('/doctors', function () {
+    return view('front.pages.doctors');
+});
+Route::get('/make-an-appointment', function () {
+    return view('front.pages.booking');
+});
+Route::get('/contact-us', function () {
+    return view('front.pages.contact');
+});
+Route::get('/login', function () {
+    return view('front.pages.login');
+});
+Route::get('/register', function () {
+    return view('front.pages.register');
+});
 
 Route::get('/dashboard', [AdminController::class, 'index'])->name('admin.index');
 
-Route::get('/majors', [MajorController::class, 'index'])->name('major.index');
-Route::get('/majors/create', [MajorController::class, 'create'])->name('major.create');
-Route::post('/majors/store', [MajorController::class, 'store'])->name('major.store');
-Route::get('/majors/edit/{id}', [MajorController::class, 'edit'])->name('major.edit');
-Route::put('/majors/update/{id}', [MajorController::class, 'update'])->name('major.update');
-Route::delete('/majors/delete/{id}', [MajorController::class, 'destroy'])->name('major.destroy');
+Route::get('/admin/majors', [MajorController::class, 'index'])->name('major.index');
+Route::get('/admin/majors/create', [MajorController::class, 'create'])->name('major.create');
+Route::post('/admin/majors/store', [MajorController::class, 'store'])->name('major.store');
+Route::get('/admin/majors/edit/{id}', [MajorController::class, 'edit'])->name('major.edit');
+Route::put('/admin/majors/update/{id}', [MajorController::class, 'update'])->name('major.update');
+Route::delete('/admin/majors/delete/{id}', [MajorController::class, 'destroy'])->name('major.destroy');
